Factor our ResponseBodyText
