+++ studio/src/pages/RequestorPage/ResponsePanel/ResponseBody/index.ts
@@ -1,1 +1,2 @@
 export * from "./ResponseBody";
+export * from "./ResponseBodyText";
